requirements.txt
watchfiles==0.19.0
rich==13.4.0
python-dateutil==2.8.2
PyYAML==6.0
